Update ORLY.pgm challenge file
Replaced the non-ascii characters from the plaintext. As they required two bytes to be correctly encoded, they shifted the flag position, meaning it was only partial at the end of the ciphertext.
This file has been validated with a deciphering scripts. It works. Just not as-is with CyberChef or the like.